+++ tremor-pipeline/src/op/qos.rs
@@ -30,13 +30,10 @@ use tremor_script::prelude::*;
 /// * we have a event delivery failure OR
 /// * we have a message triggering a circuit breaker OR
 /// * the event metadata has a `time` field that exceeds the given `timeout`
-fn is_error_insight(insight: &Event, timeout: f64) -> bool {
+fn is_error_insight(insight: &Event, timeout: u64) -> bool {
     let meta = insight.data.suffix().meta();
     meta.get("error").is_some()
         || insight.cb == CbAction::Fail
         || insight.cb == CbAction::Trigger
-        || meta
-            .get("time")
-            .and_then(Value::cast_f64)
-            .map_or(false, |v| v > timeout)
+        || meta.get_u64("time").map_or(false, |v| v > timeout)
 }